Refactored Repository methods to their correct entity repos.
