+++ API/Controllers/BookController.cs
@@ -39,7 +39,7 @@ namespace API.Controllers
             var bookTitle = string.Empty;
             if (dto.SeriesFormat == MangaFormat.Epub)
             {
-                var mangaFile = (await _unitOfWork.VolumeRepository.GetFilesForChapterAsync(chapterId)).First();
+                var mangaFile = (await _unitOfWork.ChapterRepository.GetFilesForChapterAsync(chapterId)).First();
                 using var book = await EpubReader.OpenBookAsync(mangaFile.FilePath);
                 bookTitle = book.Title;
             }
@@ -62,7 +62,7 @@ namespace API.Controllers
         [HttpGet("{chapterId}/book-resources")]
         public async Task<ActionResult> GetBookPageResources(int chapterId, [FromQuery] string file)
         {
-            var chapter = await _unitOfWork.VolumeRepository.GetChapterAsync(chapterId);
+            var chapter = await _unitOfWork.ChapterRepository.GetChapterAsync(chapterId);
             var book = await EpubReader.OpenBookAsync(chapter.Files.ElementAt(0).FilePath);
 
             var key = BookService.CleanContentKeys(file);
@@ -81,7 +81,7 @@ namespace API.Controllers
         {
             // This will return a list of mappings from ID -> pagenum. ID will be the xhtml key and pagenum will be the reading order
             // this is used to rewrite anchors in the book text so that we always load properly in FE
-            var chapter = await _unitOfWork.VolumeRepository.GetChapterAsync(chapterId);
+            var chapter = await _unitOfWork.ChapterRepository.GetChapterAsync(chapterId);
             using var book = await EpubReader.OpenBookAsync(chapter.Files.ElementAt(0).FilePath);
             var mappings = await _bookService.CreateKeyToPageMappingAsync(book);
